Add the default mark rti if bot is setup to ignore rti targets

+++ playerbot/PlayerbotFactory.cpp
@@ -1652,7 +1652,7 @@ void PlayerbotFactory::InitEquipment(bool incremental, bool syncWithMaster)
 
                     // do not use items that required level is too low compared to bot's level
                     uint32 reqLevel = sRandomItemMgr.GetMinLevelFromCache(newItemId);
-                    if (reqLevel && proto->Quality < ITEM_QUALITY_LEGENDARY && abs((int)bot->GetLevel() - (int)reqLevel) > sPlayerbotAIConfig.randomGearMaxDiff)
+                    if (reqLevel && proto->Quality < ITEM_QUALITY_LEGENDARY && abs((int)bot->GetLevel() - (int)reqLevel) > (int)sPlayerbotAIConfig.randomGearMaxDiff)
                         continue;
 
                     // filter tank weapons
@@ -1913,7 +1913,7 @@ void PlayerbotFactory::InitSecondEquipmentSet()
 
             // do not use items that required level is too low compared to bot's level
             uint32 reqLevel = sRandomItemMgr.GetMinLevelFromCache(itemId);
-            if (reqLevel && proto->Quality < ITEM_QUALITY_LEGENDARY && abs((int)bot->GetLevel() - (int)reqLevel) > sPlayerbotAIConfig.randomGearMaxDiff)
+            if (reqLevel && proto->Quality < ITEM_QUALITY_LEGENDARY && abs((int)bot->GetLevel() - (int)reqLevel) > (int)sPlayerbotAIConfig.randomGearMaxDiff)
                 continue;
 
             if (!CanEquipItem(proto, desiredQuality))
@@ -3265,7 +3265,7 @@ void PlayerbotFactory::InitInventoryEquip()
 
         // do not use items that required level is too low compared to bot's level
         uint32 reqLevel = sRandomItemMgr.GetMinLevelFromCache(itemId);
-        if (reqLevel && proto->Quality < ITEM_QUALITY_LEGENDARY && abs((int)bot->GetLevel() - (int)reqLevel) > sPlayerbotAIConfig.randomGearMaxDiff)
+        if (reqLevel && proto->Quality < ITEM_QUALITY_LEGENDARY && abs((int)bot->GetLevel() - (int)reqLevel) > (int)sPlayerbotAIConfig.randomGearMaxDiff)
             continue;
 
         if ((proto->Class != ITEM_CLASS_ARMOR && proto->Class != ITEM_CLASS_WEAPON) || (proto->Bonding == BIND_WHEN_PICKED_UP ||
